Move messengers to separate folder
