+++ app/scripts/controller-init/messengers/snaps/cronjob-controller-messenger.test.ts
@@ -1,5 +1,5 @@
 import { Messenger, RestrictedMessenger } from '@metamask/base-controller';
-import { getCronjobControllerMessenger } from './cronjob-controller-messenger';
+import { getCronjobControllerMessenger } from './cronjob-controller-messenger.ts';
 
 describe('getCronjobControllerMessenger', () => {
   it('returns a restricted controller messenger', () => {
