+++ app/scripts/controller-init/messengers/snaps/snap-insights-controller-messenger.ts
