+++ app/scripts/controller-init/messengers/snaps/snap-interface-controller-messenger.ts
