+++ app/scripts/controller-init/messengers/snaps/snaps-registry-messenger.ts
