+++ app/scripts/controller-init/snaps/index.ts
@@ -1,20 +1,7 @@
 export { CronjobControllerInit } from './cronjob-controller-init';
-export { getCronjobControllerMessenger } from './cronjob-controller-messenger';
 export { ExecutionServiceInit } from './execution-service-init';
-export { getExecutionServiceMessenger } from './execution-service-messenger';
 export { RateLimitControllerInit } from './rate-limit-controller-init';
-export {
-  getRateLimitControllerMessenger,
-  getRateLimitControllerInitMessenger,
-} from './rate-limit-controller-messenger';
 export { SnapControllerInit } from './snap-controller-init';
-export {
-  getSnapControllerMessenger,
-  getSnapControllerInitMessenger,
-} from './snap-controller-messenger';
 export { SnapInsightsControllerInit } from './snap-insights-controller-init';
-export { getSnapInsightsControllerMessenger } from './snap-insights-controller-messenger';
 export { SnapInterfaceControllerInit } from './snap-interface-controller-init';
-export { getSnapInterfaceControllerMessenger } from './snap-interface-controller-messenger';
 export { SnapsRegistryInit } from './snaps-registry-init';
-export { getSnapsRegistryMessenger } from './snaps-registry-messenger';
